+++ test/unit/utils.js
@@ -1,0 +1,10 @@
+const loggerMock = {
+    log: () => {},
+    debug: () => {},
+    info: () => {},
+    warn: () => {},
+    error: () => {},
+    silly: () => {},
+};
+
+module.exports.loggerMock = loggerMock;
